feat: add generate calcules, save result, refresh password

+++ src/domain/models/result-experiment.entity.ts
@@ -30,7 +30,7 @@ export class ResultExperiment {
   @JoinColumn({ name: 'experimentId' })
   experiment: Experiment;
 
-  @Column("int", { array: true, default: [] })
+  @Column("decimal", { array: true, default: [] })
   results: number[];
 
   @CreateDateColumn({ name: 'createdAt' })
